Add button to create a meta-issue on GitHub
This adds a button to the "How to get horizontal review" section to generate
a meta-issue to track horizontal review steps in a GitHub repository, following
a suggestion made by @anssiko in:
https://github.com/w3c/openscreenprotocol/issues/284#issuecomment-849758776

Some notes:
- Taking that approach instead of creating an issue template, as I believe that
the issue template would be seen whenever someone tries to create a new issue
and the horizontal reviews issue will only be useful once or twice per
specification.
- Sub-details sections are not included because they would grow the length of
the body beyond what can be passed to GitHub in a query string
- To keep script simple, actual steps need to be made explicit by enclosing the
prose for each step into a `<span class="step">` construct.